new(settings): [Account] Ability to delete your LunaSea account and all cloud data (#511)

+++ lib/modules/settings/routes/account/widgets/config_delete_tile.dart
@@ -2,10 +2,21 @@ import 'package:flutter/material.dart';
 import 'package:lunasea/core.dart';
 import 'package:lunasea/modules/settings.dart';
 
-class SettingsAccountDeleteConfigurationTile extends StatelessWidget {
+class SettingsAccountDeleteConfigurationTile extends StatefulWidget {
   const SettingsAccountDeleteConfigurationTile({
     Key key,
   }) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _State();
+}
+
+class _State extends State<SettingsAccountDeleteConfigurationTile> {
+  LunaLoadingState _loadingState = LunaLoadingState.INACTIVE;
+
+  void updateState(LunaLoadingState state) {
+    if (mounted) setState(() => _loadingState = state);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -14,27 +25,35 @@ class SettingsAccountDeleteConfigurationTile extends StatelessWidget {
       title: LunaText.title(text: 'settings.DeleteCloudBackup'.tr()),
       subtitle:
           LunaText.subtitle(text: 'settings.DeleteCloudBackupDescription'.tr()),
-      trailing: LunaIconButton(icon: Icons.cloud_off_rounded),
+      trailing: LunaIconButton(
+        icon: Icons.cloud_off_rounded,
+        loadingState: _loadingState,
+      ),
       onTap: () async => _delete(context),
     );
   }
 
   Future<void> _delete(BuildContext context) async {
+    if (_loadingState == LunaLoadingState.ACTIVE) return;
+    updateState(LunaLoadingState.ACTIVE);
+
     try {
       List<LunaFirebaseBackupDocument> documents =
           await LunaFirebaseFirestore().getBackupEntries();
       Tuple2<bool, LunaFirebaseBackupDocument> result =
           await SettingsDialogs().getBackupFromCloud(context, documents);
       if (result.item1) {
-        LunaFirebaseFirestore()
+        await LunaFirebaseFirestore()
             .deleteBackupEntry(result.item2.id)
             .then((_) => LunaFirebaseStorage().deleteBackup(result.item2.id))
-            .then((_) => showLunaSuccessSnackBar(
-                  title: 'settings.DeleteCloudBackupSuccess'.tr(),
-                  message: result.item2.title
-                      .replaceAll('\n', ' ${LunaUI.TEXT_EMDASH} '),
-                ))
-            .catchError((error, stack) {
+            .then((_) {
+          updateState(LunaLoadingState.INACTIVE);
+          showLunaSuccessSnackBar(
+            title: 'settings.DeleteCloudBackupSuccess'.tr(),
+            message:
+                result.item2.title.replaceAll('\n', ' ${LunaUI.TEXT_EMDASH} '),
+          );
+        }).catchError((error, stack) {
           LunaLogger().error('Firebase Backup Deletion Failed', error, stack);
           showLunaErrorSnackBar(
             title: 'settings.DeleteCloudBackupFailure'.tr(),
@@ -49,5 +68,6 @@ class SettingsAccountDeleteConfigurationTile extends StatelessWidget {
         error: error,
       );
     }
+    updateState(LunaLoadingState.INACTIVE);
   }
 }
